prefixable with string key => separate class that does not store another set of keys + semaphore
