+++ Caching.Memory/Library/Dictionary/Prefix/PrefixableMemoryCache.cs
@@ -3,7 +3,7 @@
 namespace Woody230.Caching.Memory;
 
 /// <summary>
-/// Represents a memory cache with a prefixed key. 
+/// Represents a memory cache with a prefixed key specific to the type of key.
 /// </summary>
 public abstract class PrefixableMemoryCache<TKey, TValue>: IPrefixableMemoryCache<TKey, TValue>
 {
@@ -56,4 +56,44 @@ public abstract class PrefixableMemoryCache<TKey, TValue>: IPrefixableMemoryCach
     protected abstract string GetPrefix(TKey key);
 
     private string GetPrefixedKey(TKey key) => GetPrefix(key) + key;
+}
+
+/// <summary>
+/// Represents a memory cache with a prefixed key. 
+/// </summary>
+public abstract class PrefixableMemoryCache<TValue> : IPrefixableMemoryCache<string, TValue>
+{
+    private readonly IGenericMemoryCache<string, TValue> _cache;
+
+    public PrefixableMemoryCache(IGenericMemoryCache<string, TValue> cache)
+    {
+        _cache = cache;
+    }
+
+    public IEnumerable<string> Keys => _cache.Keys.Select(GetPrefixedKey);
+
+    public void Remove(string key)
+    {
+        var prefixedKey = GetPrefixedKey(key);
+        _cache.Remove(prefixedKey);
+    }
+
+    public void Set(string key, TValue value, IMemoryCacheEntryOptions options)
+    {
+        var prefixedKey = GetPrefixedKey(key);
+        _cache.Set(prefixedKey, value, options);
+    }
+
+    public bool TryGetValue(string key, [MaybeNullWhen(false)] out TValue value)
+    {
+        var prefixedKey = GetPrefixedKey(key);
+        return _cache.TryGetValue(prefixedKey, out value);
+    }
+
+    /// <summary>
+    /// Gets the prefix associated with the <paramref name="key"/>.
+    /// </summary>
+    protected abstract string GetPrefix(string key);
+
+    private string GetPrefixedKey(string key) => GetPrefix(key) + key;
 }
